Fix building on gcc
There were several places where I don't know how it was building before,
but it certainly wasn't on my machine (Debian). I've fixed a few
simple syntax issues, and a struct member name that was being used
incorrectly.
In addition to this, I have also "fixed" the reading of the data window
and the display window of OpenEXR files. Without being very familiar
with this code base or the OpenEXR file format (I had a brief look
to figure out what the display window and data window were, but
not much else) I am nervous about making this change. If anyone can
test OpenEXR decoding, or instruct me on how to test them, that
would probably be a good idea.