Allow building with OpenSSL 1.10
Apparently the newer version of OpenSSL that I have is not
building with the current libasdcp because there were some source
incompatible changes. The master version of libasdcp will probably
fix this at some point, but for anyone needing a version
of libasdcp that compiles with the new OpenSSL, this branch
will have the fix.